infralocationanalysis/requirements.txt
async-generator==1.10
attrs==22.1.0
autopep8==1.7.0
beautifulsoup4==4.11.1
browsermob-proxy==0.8.0
bs4==0.0.1
charset-normalizer
contextlib2==21.6.0
decorator==5.1.1
dnspython==2.0.0
docopt==0.6.2
h11==0.13.0
idna==3.4
ipwhois==1.2.0
outcome==1.1.0
psutil==5.9.2
pycodestyle==2.9.1
PySocks==1.7.1
python-dotenv==0.20.0
requests
requests-file==1.5.1
schema==0.7.5
selenium==4.0.0a7
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
tldextract==3.1.2
toml==0.10.2
tqdm==4.64.1
trio==0.19.0
trio-websocket==0.9.2
urllib3==1.26.12
validators==0.20.0
webdriver-manager==3.7.1
wsproto==1.0.0
mysql-connector==2.2.9
mysql-connector-python==8.0.32
matplotlib==3.7.1
numpy==1.24.2
scipy==1.9.1
psycopg2-binary==2.9.6